Various cleanups based on compiler warnings.
git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@3315 e44c5d27-f830-0410-8ce8-974f7a764be6

+++ src/fvm/fvm_box.c
@@ -781,10 +781,10 @@ fvm_box_set_dump(const fvm_box_set_t  *boxes,
       if (bmin[j] > bmax[j])
         bft_error(__FILE__, __LINE__, 0,
                   _("Inconsistent box found (min > max):\n"
-                    "  global number:  %u\n"
+                    "  global number:  %llu\n"
                     "  min       :  %10.4g\n"
                     "  max       :  %10.4g\n"),
-                  boxes->g_num[i], bmin[j], bmax[j]);
+                  (unsigned long long)(boxes->g_num[i]), bmin[j], bmax[j]);
     }
   }
